requirements.txt
streamlit
apify-client
groq
instaloader
python-dotenv
pandas
requests
pytest
